TCP requests have the NameNodeResult objects sent back directly to the clients without converting to Json. Should hopefully save time/memory.

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/serverless/cache/FunctionMetadataMap.java
@@ -121,7 +121,7 @@ public class FunctionMetadataMap {
      * @param overwriteExisting Overwrite an existing entry.
      * @return `true` if entry was added to the cache, otherwise `false`.
      */
-    public boolean addEntry(String path, int function, boolean overwriteExisting) {
+    public boolean addEntry(String path, long function, boolean overwriteExisting) {
         String pathToCache = getPathToCache(path);
 
         if (LOG.isDebugEnabled())
